refactor: move magic_link endpoints to dedicated api file
